Rework final version (XSIMD) requires further testing

+++ mudock/include/mudock/type_alias.hpp
@@ -2,6 +2,6 @@
 
 namespace mudock {
 
-  using fp_type = double;
+  using fp_type = float;
 
 } // namespace mudock
